moving tests to correct folders to avoid circular dependencies
