+++ go/common/rlp_test.go
@@ -1,98 +1,31 @@
 package common
 
 import (
-	common2 "github.com/otherview/obscuro-playground/go/obscuro-node/common"
-	"github.com/otherview/obscuro-playground/integration/wallet-mock"
-	"testing"
-	"time"
-
 	"github.com/ethereum/go-ethereum/rlp"
 	"github.com/google/uuid"
-	"github.com/otherview/obscuro-playground/obscuro/common"
+	"testing"
 )
 
-func TestSerialiseL2Tx(t *testing.T) {
-	tx := common.L2Tx{
-		Id:     uuid.New(),
-		TxType: common.TransferTx,
-		Amount: 100,
-		From:   wallet_mock.New().Address,
-		To:     wallet_mock.New().Address,
-	}
-	bytes, err := rlp.EncodeToBytes(tx)
-	if err != nil {
-		panic(err)
-	}
-	tx1 := common.L2Tx{}
-	err2 := rlp.DecodeBytes(bytes, &tx1)
-	if err2 != nil {
-		panic(err2)
-	}
-	if tx1.Id != tx.Id {
-		t.Errorf("tx deserialized incorrectly\n")
-	}
-}
-
-func TestSerialiseRollup(t *testing.T) {
-	tx := common.L2Tx{
-		Id:     uuid.New(),
-		TxType: common.TransferTx,
-		Amount: 100,
-		From:   wallet_mock.New().Address,
-		To:     wallet_mock.New().Address,
-	}
-	rollup := common2.Rollup{
-		Height:       1,
-		RootHash:     uuid.New(),
-		Agg:          1,
-		ParentHash:   uuid.New(),
-		CreationTime: time.Now(),
-		L1Proof:      uuid.New(),
-		Nonce:        100,
-		State:        "",
-		Withdrawals:  nil,
-		Transactions: []common.L2Tx{tx},
-	}
-	_, read, err := rlp.EncodeToReader(&rollup)
-	if err != nil {
-		panic(err)
-	}
-	r1 := common2.Rollup{}
-
-	err2 := rlp.Decode(read, &r1)
-
-	if err2 != nil {
-		panic(err2)
-	}
-	if r1.Hash() != rollup.Hash() {
-		t.Errorf("rollup deserialized incorrectly\n")
-	}
-	if r1.Transactions[0].Id != rollup.Transactions[0].Id {
-		t.Errorf("rollup deserialized incorrectly\n")
-	}
-}
 
 func TestSerialiseBlock(t *testing.T) {
-	tx := L1Tx{
+	tx := &L1Tx{
 		Id:     uuid.New(),
 		TxType: DepositTx,
 		Amount: 100,
-		Dest:   wallet_mock.New().Address,
+		//Dest:   wallet_mock.New().Address,
 	}
 	block := Block{
-		Height:       1,
-		RootHash:     uuid.New(),
-		Miner:        1,
-		ParentHash:   uuid.New(),
-		ReceiveTime:  time.Now(),
-		Nonce:        100,
-		Transactions: []L1Tx{tx},
+		Header:       GenesisBlock.Header,
+		Transactions: Transactions{tx},
+		hash:         GenesisBlock.hash,
+		height:       GenesisBlock.height,
+		size:         GenesisBlock.size,
 	}
 	bytes, err := rlp.EncodeToBytes(block)
 	if err != nil {
 		panic(err)
 	}
-	b1 := Block{Transactions: make([]L1Tx, 1)}
+	b1 := Block{Transactions: Transactions{tx}}
 	err2 := rlp.DecodeBytes(bytes, &b1)
 	if err2 != nil {
 		panic(err2)
@@ -106,26 +39,24 @@ func TestSerialiseBlock(t *testing.T) {
 }
 
 func TestPlay(t *testing.T) {
-	tx := L1Tx{
+	tx := &L1Tx{
 		Id:     uuid.New(),
 		TxType: DepositTx,
 		Amount: 100,
-		Dest:   wallet_mock.New().Address,
+		//Dest:   wallet_mock.New().Address,
 	}
 	block := Block{
-		Height:       1,
-		RootHash:     uuid.New(),
-		Miner:        1,
-		ParentHash:   uuid.New(),
-		ReceiveTime:  time.Now(),
-		Nonce:        100,
-		Transactions: []L1Tx{tx},
+		Header:       GenesisBlock.Header,
+		Transactions: Transactions{tx},
+		hash:         GenesisBlock.hash,
+		height:       GenesisBlock.height,
+		size:         GenesisBlock.size,
 	}
 	bytes, err := rlp.EncodeToBytes(block)
 	if err != nil {
 		panic(err)
 	}
-	b1 := Block{Transactions: make([]L1Tx, 1)}
+	b1 := Block{Transactions: Transactions{tx}}
 	err2 := rlp.DecodeBytes(bytes, &b1)
 	if err2 != nil {
 		panic(err2)
