+++ go/obscuro-node/common/types_test.go
@@ -1,0 +1,2 @@
+package common
+
